requirements.txt
amqp==5.0.6
beautifulsoup4==4.10.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
comtypes==1.1.10
docopt==0.6.2
Flask==2.0.2
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kit==0.2.15
kombu==5.2.2
MarkupSafe==2.0.1
MouseInfo==0.1.3
Pillow==8.4.0
prompt-toolkit==3.0.24
PyAudio @ file:///D:/Programming%20Train/jarvis/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
python-decouple==3.5
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.3
pywhatkit==5.2
pywin32==302
PyYAML==6.0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.28
urllib3==1.26.7
utils==1.0.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
wikipedia==1.4.0
